BE/Enterprise_NT/src/main/java/net/cltech/enterprisent/domain/operation/pathology/Audio.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package net.cltech.enterprisent.domain.operation.pathology;

import com.fasterxml.jackson.annotation.JsonInclude;
import net.cltech.enterprisent.domain.masters.common.PathologyAudit;
import org.jsondoc.core.annotation.ApiObject;
import org.jsondoc.core.annotation.ApiObjectField;

/**
 * Clase que representa la informacion de los audios de Patologia
 *
 * @version 1.0.0
 * @author omendez
 * @since 16/06/2021
 * @see Creación
 */
@ApiObject(
        group = "Patología",
        name = "Audio",
        description = "Muestra informacion de los audios de patologia que usa el API"
)
@JsonInclude(JsonInclude.Include.NON_NULL)
public class Audio extends PathologyAudit
{
    @ApiObjectField(name = "id", description = "Id", required = true, order = 1)
    private Integer id;
    @ApiObjectField(name = "name", description = "Nombre", required = true, order = 2)
    private String name;
    @ApiObjectField(name = "extension", description = "Extension", required = true, order = 3)
    private String extension;
    @ApiObjectField(name = "url", description = "URL", required = true, order = 4)
    private String url;

    public Integer getId() {
        return id;
    }

    public void setId(Integer id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getExtension() {
        return extension;
    }

    public void setExtension(String extension) {
        this.extension = extension;
    }

    public String getUrl() {
        return url;
    }

    public void setUrl(String url) {
        this.url = url;
    }
}
